Phase 8: perspective cube with depth-correct rendering
- Cube mesh (24 verts, 36 indices, 6 coloured faces) centred at origin in
  model space; placed at z=-3 via Translation*RotateY(45)*RotateX(30) model
  matrix to show three faces simultaneously.
- SetPerspectiveProjection (FOV=60, 4:3, near=0.1, far=100) added to DC
  backend.
- Index pipeline: mpBoundIndices promoted to uint32*; SetIndexData(CIndexBuffer*)
  now binds mesh index buffer directly; SetIndexData(CIndexBufferChunk) promotes
  uint16→uint32 via scratch buffer.
- RenderPrimitives: switch to DEPTHCMP_GEQUAL (1/w depth, larger=closer) —
  confirmed working now that faces have distinct depths under perspective.
- gen_modl.py: cube generator added alongside triangle; cube placed at origin.

+++ bp/Source/Renderer/Base/Backend/DC/DCCRenderBackend.h
@@ -313,7 +313,7 @@ private:
    // Bound geometry state — set by SetVertexData / SetIndexData, consumed by RenderPrimitives.
    const CVertexData *         mpBoundVertexData;
    CShaderVertexDataBinding    mBoundBinding;
-   const uint16 *              mpBoundIndices;
+   const uint32 *              mpBoundIndices;
 
    bool               mDepthWriteEnabled;
    bool               mBlendEnabled;
